feat: impl a CliHostTree structure that implement to_string
