+++ src/commands/service/mod.rs
@@ -4,7 +4,6 @@ mod remove;
 mod restore;
 mod start;
 mod stop;
-mod tree;
 
 pub use apply::apply;
 pub use new::new;
@@ -12,4 +11,3 @@ pub use remove::remove;
 pub use restore::restore;
 pub use start::start;
 pub use stop::stop;
-pub use tree::display_tree;
